python-service/requirements.txt
fastapi
uvicorn
torch
open_clip_torch
Pillow
python-multipart
faiss-cpu
httpx